+++ inc/c_abi/Tecs_entity.h
@@ -4,11 +4,14 @@
 
 #ifdef __cplusplus
 extern "C" {
+    #include <stdint.h>
 #else
+    #include <stdint.h>
+
 typedef uint8_t bool;
 #endif
 
-#include <stdint.h>
+#include <stddef.h>
 
 typedef void TecsLock;
 typedef uint64_t TecsEntity;
@@ -17,8 +20,8 @@ TECS_EXPORT bool Tecs_entity_exists(TecsLock *dynLockPtr, TecsEntity entity);
 TECS_EXPORT bool Tecs_entity_existed(TecsLock *dynLockPtr, TecsEntity entity);
 TECS_EXPORT bool Tecs_entity_has(TecsLock *dynLockPtr, TecsEntity entity, size_t componentIndex);
 TECS_EXPORT bool Tecs_entity_had(TecsLock *dynLockPtr, TecsEntity entity, size_t componentIndex);
-TECS_EXPORT bool Tecs_entity_has_bitset(TecsLock *dynLockPtr, TecsEntity entity, unsigned long long componentBits);
-TECS_EXPORT bool Tecs_entity_had_bitset(TecsLock *dynLockPtr, TecsEntity entity, unsigned long long componentBits);
+TECS_EXPORT bool Tecs_entity_has_bitset(TecsLock *dynLockPtr, TecsEntity entity, uint64_t componentBits);
+TECS_EXPORT bool Tecs_entity_had_bitset(TecsLock *dynLockPtr, TecsEntity entity, uint64_t componentBits);
 TECS_EXPORT const void *Tecs_entity_const_get(TecsLock *dynLockPtr, TecsEntity entity, size_t componentIndex);
 TECS_EXPORT const void *Tecs_const_get_entity_storage(TecsLock *dynLockPtr, size_t componentIndex);
 TECS_EXPORT void *Tecs_entity_get(TecsLock *dynLockPtr, TecsEntity entity, size_t componentIndex);
